requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.66.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.11.0
highway-env==1.9.1
idna==3.8
imageio==2.35.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
importlib_resources==6.4.0
jax-jumpy==1.0.0
Jinja2==3.1.4
joblib==0.14.1
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mujoco-py==2.1.2.14
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygame==2.1.0
pyglet==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rpm @ file:///home/conda/feedstock_root/build_artifacts/rpm-tools_1594748004177/work/python
rsa==4.9
-e git+https://github.com/geraudnt/skill_machines.git@72cba7765c398820e9e400d066acdf3c5e2ce5a0#egg=safety_gym&subdirectory=envs/safety_gym/safety-gym
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stable-baselines3==2.0.0
sympy==1.13.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.6
Werkzeug==3.0.4
wrapt==1.17.0
xmltodict==0.13.0
zipp==3.20.0
zmq==0.0.0
